+++ app/Http/Controllers/SeriesController.php
@@ -15,6 +15,8 @@ class SeriesController extends Controller
      */
     public function index()
     {
+        $this->authorize('view_list', Series::class);
+
         return view('pages.series.index');
     }
 
@@ -25,6 +27,8 @@ class SeriesController extends Controller
      */
     public function create()
     {
+        $this->authorize('view_create', Series::class);
+        
         return view('pages.series.create');
     }
 
@@ -36,6 +40,8 @@ class SeriesController extends Controller
      */
     public function store(SeriesStoreRequest $request)
     {
+        $this->authorize('create', Series::class);
+        
         $series = Series::create($request->all());
         if($request->hasFile('cover')){
             $originalExtension = $request->file('cover')->getClientOriginalExtension();
@@ -66,6 +72,8 @@ class SeriesController extends Controller
      */
     public function edit(Series $series)
     {
+        $this->authorize('view_edit', Series::class);
+        
         return view('pages.series.edit', [
             'series' => $series
         ]);
@@ -80,6 +88,8 @@ class SeriesController extends Controller
      */
     public function update(SeriesStoreRequest $request, Series $series)
     {
+        $this->authorize('update', Series::class);
+        
         $series->update($request->all());
         if ($request->hasFile('cover')) {
             $originalExtension = $request->file('cover')->getClientOriginalExtension();
@@ -104,6 +114,8 @@ class SeriesController extends Controller
      */
     public function destroy(Series $series)
     {
+        $this->authorize('destroy', Series::class);
+        
         $series->delete();
         return redirect(route('series.index'));
     }
